+++ packages/better_player/lib/src/asms/better_player_asms_utils.dart
@@ -50,7 +50,7 @@ class BetterPlayerAsmsUtils {
       final response = await request.close();
       return await response.transform(const Utf8Decoder()).join();
     } catch (exception) {
-      BPLog.error(
+      Logger.error(
         'GetDataFromUrl failed: $exception',
         error: exception,
         breadcrumb: 'AsmsUtils',