+++ packages/better_player/lib/src/core/better_player_logger.dart
@@ -104,7 +104,4 @@ class BetterPlayerLogger {
 }
 
 /// Shorthand for [BetterPlayerLogger].
-typedef BPLog = BetterPlayerLogger;
-
-/// Shorthand for [BetterPlayerLogger].
 typedef Logger = BetterPlayerLogger;